refactor components directory naming for consistency-convention
