+++ src/client/components/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 
-import Home from './home/Home';
+import Home from './Home/Home';
 
 const App = () => (
   <BrowserRouter>
