+++ src/client/components/Cell/Cell.js
