+++ src/client/components/Label/Label.js
