+++ src/client/components/Label/Label.scss
