+++ src/client/components/Player/Player.scss
