+++ src/client/components/Row/Row.js
@@ -1,15 +1,15 @@
 import React from 'react';
-import Cell from '../cell/Cell';
+import Cell from '../Cell/Cell';
 import { heightSize } from '../../constants';
 
 import './Row.scss';
 
 const Row = ({ row }) => {
   let cells = [];
-  // Game grid.
+  // Game Grid.
   if (typeof row !== 'number') {
     cells = row.map((e, i) => <Cell key={i} fill={e.fill} color={e.color} />);
-  // Player spectrum.
+  // Player Spectrum.
   } else {
     for (let i = 0; i < heightSize; i++) {
       cells.push(<Cell key={i} fill={row <= i} color="grey" />);
