+++ src/client/components/Spectrum/Spectrum.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Row from '../row/Row';
+import Row from '../Row/Row';
 
 import './Spectrum.scss';
