+++ src/client/components/Spectrum/Spectrum.scss
